+++ src/cli.rs
@@ -23,7 +23,7 @@ use crate::{
 #[command(propagate_version = true, infer_subcommands = true)]
 pub struct Cli {
     #[command(subcommand)]
-    pub command: HimalayaCommand,
+    pub command: Option<HimalayaCommand>,
 
     /// Override the default configuration file path
     ///